requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bokeh==3.3.4
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.0
dask==2024.1.0
distributed==2024.1.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.104.1
fsspec==2023.12.2
greenlet==3.0.3
h11==0.14.0
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
locket==1.0.0
lz4==4.3.3
MarkupSafe==2.1.4
msgpack==1.0.7
numpy==1.26.3
openpyxl==3.1.2
packaging==23.2
pandas==2.1.2
partd==1.4.1
pillow==10.2.0
psutil==5.9.8
pyarrow==15.0.0
pyarrow-hotfix==0.6
pydantic==2.5.3
pydantic_core==2.14.6
pyodbc==5.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.23
starlette==0.27.0
tblib==3.0.0
toolz==0.12.1
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.23.2
xyzservices==2023.10.1
zict==3.0.0
zipp==3.17.0
psycopg2-binary
geojson 